+++ HISTORY.md
@@ -6,7 +6,11 @@
 * Fixes bug where `hybracter` would crash if the input long reads were not gzipped #51 thanks @wanyuac.
 * Logic changes for chromosome contigs and circularity. If hybracter assemblies a contig that is greater than the minimum chromosome length but not marked as circular by Flye, this will now be denoted as a chromosome, but not circular. It will be polished and in the final `_chromosome.fasta` output and it will not be rotated by `dnaapler`. 
   * These were previously being excluded, which was missing chromosomes with structural heterogeneity (causing the chromosome not to completely circularise) or bacteria with linear chromosomes like [_Borrelia_](https://www.nature.com/articles/37551).
-* Logic changes for short read polishing. Logic added to run Polypolish `--careful` and skip pypolca if the SR coverage estimate is below 5x (note: FASTA files for pypolca will be generated in the processing directory to play nice with Snakemake, but these will be identical to the polypolish output). For 5-25x coverage, `polypolish --careful` and `pypolca --careful` will be run. For >25x coverage, `polypolish` default and `pypolca --careful` will be run. 
+* Changes to short read polishing. Logic added to run `polypolish` v0.6.0 with `--careful` and skip pypolca if the SR coverage estimate is below 5x (note: FASTA files for pypolca will be generated in the processing directory to play nice with Snakemake, but these will be identical to the polypolish output). 
+* For 5-25x coverage, `polypolish --careful` and `pypolca` with `--careful` will be run. For >25x coverage, `polypolish` default and `pypolca` with `--careful` will be run. 
+* By default, `--logic` defaults to `last` for `hybracter hybrid`, as there we have found that the polishing strategy implemented above never makes the assembly worse (a pre-print is coming!). We suggest never using `--logic best` with `hybracter hybrid`.
+
+
 
 ## v0.6.0 (18 January 2024)